Added simple db connection, startup will fail when connection does not succeed.
Default is to connect to postgresql on localhost:5432, with username zophrenic, empty password and database zophrenic.

You can change this with the environment vars:
 ZP_DBHOST, ZP_DBPORT, ZP_DBUSER, ZP_DBPASSWORD and ZP_DB

Changed webmachine env names to ZP_IP and ZP_PORT.

Added utility functions to pgsql:
- assoc/2 and assoc/3 to fetch a property list of all rows
- equery1/2, equery1/3 and squery1/2 to fetch a single value from a query
- last_id/2 to return the last insert id for a table (and the current session).

Removed the application behaviour from epgsql_pool, it is now only a supervisor in the supervisor tree of the zophrenic application.


git-svn-id: https://Neo:8443/svn/zotonic/trunk@143 39ac7d6a-fe55-4cd0-aea4-b13bf11f1711