Added options to inline images, and a proper template.
For now the only option is the alignment of the image, but this can be
expanded easily. Options are json-encoded inside the HTML comment
marker, and are json-decoded by the erlydtl modifier.

The inline images are now served from a template, _body_media.tpl,
which can be overruled for a project, etc.



git-svn-id: https://Neo:8443/svn/zotonic/trunk@1220 39ac7d6a-fe55-4cd0-aea4-b13bf11f1711