Added z_datamodel module, for easy, per-module installation of parts of a zotonic datamodel.
Still experimental.

Module can install predicates and categories, and sets a flag on them
so it knows that the calling module has installed those. On a module
install, that part of the datamodel will be checked agains the curent
definition and updated accordingly.

On a module deinstall, the datamodel should be removed (if there are
no resources tied to them).



git-svn-id: https://Neo:8443/svn/zotonic/trunk@1336 39ac7d6a-fe55-4cd0-aea4-b13bf11f1711